notification-service/requirements.txt
django==5.1.4
djangorestframework==3.15.2
psycopg2-binary==2.9.9
pytest==9.0.3
pytest-django==4.12.0
celery==5.3.6
redis==5.0.1
